+++ Modules/People/Resources/views/bulkSearch.blade.php
@@ -6,6 +6,7 @@
 
     {!! Form::open(['route' => 'people.doBulkSearch']) !!}
         {{ Form::bsTextarea('data', null, ['autofocus'], __('app.data')) }}
+        {{ Form::bsSelect('order', $orders, $order, [], __('app.order')) }}
         <p>
             {{ Form::bsSubmitButton(__('app.search'), 'search') }}
         </p>